+++ src/mol-gl/shader/chunks/check-picking-alpha.glsl.ts
@@ -1,9 +1,14 @@
 export const check_picking_alpha = `
 float viewZ = depthToViewZ(uIsOrtho, fragmentDepth, uNear, uFar);
 float fogFactor = smoothstep(uFogNear, uFogFar, abs(viewZ));
-float alpha = (1.0 - fogFactor) * uAlpha;
+float fogAlpha = (1.0 - fogFactor) * uAlpha;
+float alpha = uAlpha;
+#ifdef dXrayShaded
+    // add bias to make picking xray shaded elements easier
+    alpha = calcXrayShadedAlpha(alpha, normal) + (0.3 * uPickingAlphaThreshold);
+#endif
 // if not opaque enough ignore so the element below can be picked
-if (uAlpha < uPickingAlphaThreshold || alpha < 0.1) {
+if (alpha < uPickingAlphaThreshold || fogAlpha < 0.1) {
     #ifdef dTransparentBackfaces_opaque
         if (!interior) discard;
     #else