+++ src/mol-model/structure/structure/unit/bonds/data.ts
@@ -12,26 +12,26 @@ import { StructureElement } from '../../element';
 import { Bond } from '../bonds';
 import { InterUnitGraph } from '../../../../../mol-math/graph/inter-unit-graph';
 
-type IntraUnitBondProps = {
+export type IntraUnitBondProps = {
     /** Can remap even with `dynamicBonds` on, e.g., for water molecules */
     readonly canRemap?: boolean
     /** Can be cached in `ElementSetIntraBondCache` */
     readonly cacheable?: boolean
 }
 
-type IntraUnitBonds = IntAdjacencyGraph<StructureElement.UnitIndex, {
+export type IntraUnitBonds = IntAdjacencyGraph<StructureElement.UnitIndex, {
     readonly order: ArrayLike<number>,
     readonly flags: ArrayLike<BondType.Flag>
     readonly key: ArrayLike<number>,
 }, IntraUnitBondProps>
 
-namespace IntraUnitBonds {
+export namespace IntraUnitBonds {
     export const Empty: IntraUnitBonds = IntAdjacencyGraph.create([], [], [], 0, { flags: [], order: [], key: [] });
 }
 
-type InterUnitEdgeProps = { readonly order: number, readonly flag: BondType.Flag, readonly key: number }
+export type InterUnitEdgeProps = { readonly order: number, readonly flag: BondType.Flag, readonly key: number }
 
-class InterUnitBonds extends InterUnitGraph<number, StructureElement.UnitIndex, InterUnitEdgeProps> {
+export class InterUnitBonds extends InterUnitGraph<number, StructureElement.UnitIndex, InterUnitEdgeProps> {
     /** Get inter-unit bond given a bond-location */
     getBondFromLocation(l: Bond.Location) {
         return Unit.isAtomic(l.aUnit) && Unit.isAtomic(l.bUnit) ? this.getEdge(l.aIndex, l.aUnit.id, l.bIndex, l.bUnit.id) : undefined;
@@ -43,9 +43,7 @@ class InterUnitBonds extends InterUnitGraph<number, StructureElement.UnitIndex, 
     }
 }
 
-namespace InterUnitBonds {
+export namespace InterUnitBonds {
     export class UnitPairBonds extends InterUnitGraph.UnitPairEdges<number, StructureElement.UnitIndex, InterUnitEdgeProps> {}
     export type BondInfo = InterUnitGraph.EdgeInfo<StructureElement.UnitIndex, InterUnitEdgeProps>
 }
-
-export { IntraUnitBonds, IntraUnitBondProps, InterUnitBonds, InterUnitEdgeProps };