+++ src/mol-geo/geometry/image/image.ts
@@ -71,6 +71,8 @@ interface Image {
 
     /** Bounding sphere of the image */
     boundingSphere: Sphere3D
+
+    setBoundingSphere(boundingSphere: Sphere3D): void
 }
 
 namespace Image {
@@ -126,6 +128,10 @@ namespace Image {
                     currentHash = newHash;
                 }
                 return boundingSphere;
+            },
+            setBoundingSphere(sphere: Sphere3D) {
+                Sphere3D.copy(boundingSphere, sphere);
+                currentHash = hashCode(image);
             },
         };
         return image;