+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add `mol-utils/camera.ts` with `fovAdjustedPosition` and `fovNormalizedCameraPosition`
 - Show FOV normalized position in `CameraInfo` UI and use it in "Copy MVS State"
 - Support static resources in `AssetManager`
-
+- Support `pickingAlphaThreshold` when `xrayShaded` is enabled
 
 ## [v4.17.0] - 2025-05-22
 - Remove `xhr2` dependency for NodeJS, use `fetch`