+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Note that since we don't clearly distinguish between a public and private interf
   - Use `isolatedModules` tsconfig flag
   - Fix TurboPack build when using ES6 modules
 - Support `pickingAlphaThreshold` when `xrayShaded` is enabled
-- Refactor SASS to not use `@import` (fixes deprecation warnings)
+- Refactor SCSS to not use `@import` (fixes deprecation warnings)
 
 ## [v4.17.0] - 2025-05-22
 - Remove `xhr2` dependency for NodeJS, use `fetch`