+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file, following t
 Note that since we don't clearly distinguish between a public and private interfaces there will be changes in non-major versions that are potentially breaking. If we make breaking changes to less used interfaces we will highlight it in here.
 
 ## [Unreleased]
+
+## [v4.18.0] - 2025-06-08
 - MolViewSpec extension:
   - Support for label_comp_id and auth_comp_id in annotations
   - Geometric primitives - do not render if position refers to empty substructure